+++ fatiando/gravmag/tests/test_transform.py
@@ -1,4 +1,5 @@
 from __future__ import division
+import pytest
 import numpy as np
 import numpy.testing as npt
 from fatiando.gravmag import transform, prism
@@ -58,6 +59,18 @@ def test_upcontinue():
     assert np.all(check), \
         'Failed for tf (mismatch {:.2f}%)'.format(
         100*(check.size - check.sum())/check.size)
+
+
+def test_upcontinue_warning():
+    "gravmag.transform upward continuation raises warning if height <= 0"
+    model = [Prism(-1000, 1000, -500, 500, 0, 1000, {'density': 1000})]
+    shape = (100, 100)
+    x, y, z = gridder.regular([-5000, 5000, -5000, 5000], shape, z=-500)
+    data = prism.gz(x, y, z, model)
+    with pytest.warns(UserWarning):
+        up = transform.upcontinue(x, y, data, shape, height=0)
+    with pytest.warns(UserWarning):
+        up = transform.upcontinue(x, y, data, shape, height=-100)
 
 
 def test_second_horizontal_derivatives_fd():