+++ doc/changelog.rst
@@ -10,6 +10,10 @@ Version 0.5
 
 **Changes**:
 
+* Change the behavior of ``gravmag.transform.upcontinue``. Instead of raise
+  an error when 'height' <= 0 it's now warn users that in this case the
+  computation is unstable.
+  (`PR 337 <https://github.com/fatiando/fatiando/pull/337>`__)
 * Add functions ``power_density_spectra`` and ``radial_average_spectrum`` to
   ``fatiando.gravmag.transform`` to calculate the radial average power density
   spectrum of gridded potential field data.